test(otel): move tracestate fixtures to test utils
